Terminandas implementaciones a nivel básico y algunas correcciones de tipos en otras clases

+++ src/main/java/Repository/StatusRepository.java
@@ -4,5 +4,5 @@ import Beans.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StatusRepository extends JpaRepository<Status, Long> {
-    Status findByMotorcycleId (Integer id);
+    Status findByMotorcycleId (Long id);
 }
